requirements.txt
# Data Processing
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
joblib

# Visualization
matplotlib
seaborn

# env management
python-dotenv

# MLOPs
dvc==3.51.1
mlflow==2.17.0
dagshub

# Web Framework / API
uvicorn
fastapi==0.115.5

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.36

# Machine Learning / NLP
sentence-transformers==5.1.1
transformers==4.56.2
torch==2.8.0
faiss-cpu==1.7.4
accelerate
safetensors
bitsandbytes